LayoutTests/fast/profiler/call-nodelist-as-function-expected.txt

CONSOLE MESSAGE: line 59: Call NodeList as function
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) call-nodelist-as-function.html 8
CONSOLE MESSAGE: line 75:   (program) call-nodelist-as-function.html 4
CONSOLE MESSAGE: line 75:   onload call-nodelist-as-function.html 20
CONSOLE MESSAGE: line 75:    testStart call-nodelist-as-function.html 10
CONSOLE MESSAGE: line 75:     getElementsByTagName  0
CONSOLE MESSAGE: line 75:     (NodeList object)  0
CONSOLE MESSAGE: line 75:     endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
This page calls a NodeList as a function (e.g., list(0)). 

To run this test manually, load it in a debug build in the browser. You should see a (NodeList object) entry in the profile.
